Fix bugs and refactor the sessions module
Well, I forgot that `_.compact` returned an array and was using as if it
was a function I'm used to having on production environments
(`compactObj`) and that took quite a while to debug. Also, for some
reason I wasn't using the already written `sessions.load` method on the
`sessions.save` implementation, which led to code duplication in the
most ridiculous of ways. Things work fine now.